+++ app/ems/__init__.py
@@ -1,4 +1,5 @@
 from .controller import EMSController
+from .ops import WriteOp, WriteResult, WriteTarget
 from .state import StateProxy
 
-__all__ = ["EMSController", "StateProxy"]
+__all__ = ["EMSController", "StateProxy", "WriteOp", "WriteResult", "WriteTarget"]